+++ src/Oro/Component/ChainProcessor/ParameterBag.php
@@ -7,8 +7,14 @@ namespace Oro\Component\ChainProcessor;
  */
 class ParameterBag extends AbstractParameterBag
 {
-    /** @var array */
-    protected $items = [];
+    /** @var array [key => value, ...] */
+    private $items = [];
+
+    /** @var ParameterValueResolverInterface[] [key => ParameterValueResolverInterface, ...] */
+    private $resolvers = [];
+
+    /** @var array [key => value, ...] */
+    private $resolvedItems = [];
 
     /**
      * {@inheritdoc}
@@ -23,7 +29,24 @@ class ParameterBag extends AbstractParameterBag
      */
     public function get($key)
     {
-        return $this->items[$key] ?? null;
+        if (!\array_key_exists($key, $this->items)) {
+            return null;
+        }
+
+        if (\array_key_exists($key, $this->resolvedItems)) {
+            return $this->resolvedItems[$key];
+        }
+
+        $value = $this->items[$key];
+        if (isset($this->resolvers[$key])) {
+            $resolver = $this->resolvers[$key];
+            if ($resolver->supports($value)) {
+                $value = $resolver->resolve($value);
+            }
+            $this->resolvedItems[$key] = $value;
+        }
+
+        return $value;
     }
 
     /**
@@ -32,6 +55,19 @@ class ParameterBag extends AbstractParameterBag
     public function set($key, $value)
     {
         $this->items[$key] = $value;
+        unset($this->resolvedItems[$key]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setResolver($key, ?ParameterValueResolverInterface $resolver)
+    {
+        if (null === $resolver) {
+            unset($this->resolvers[$key]);
+        } else {
+            $this->resolvers[$key] = $resolver;
+        }
     }
 
     /**
@@ -39,7 +75,7 @@ class ParameterBag extends AbstractParameterBag
      */
     public function remove($key)
     {
-        unset($this->items[$key]);
+        unset($this->items[$key], $this->resolvedItems[$key]);
     }
 
     /**
@@ -47,7 +83,22 @@ class ParameterBag extends AbstractParameterBag
      */
     public function toArray()
     {
-        return $this->items;
+        $result = [];
+        foreach ($this->items as $key => $value) {
+            if (\array_key_exists($key, $this->resolvedItems)) {
+                $value = $this->resolvedItems[$key];
+            } elseif (isset($this->resolvers[$key])) {
+                $resolver = $this->resolvers[$key];
+                if ($resolver->supports($value)) {
+                    $value = $resolver->resolve($value);
+                }
+                $this->resolvedItems[$key] = $value;
+            }
+
+            $result[$key] = $value;
+        }
+
+        return $result;
     }
 
     /**
@@ -56,5 +107,14 @@ class ParameterBag extends AbstractParameterBag
     public function clear()
     {
         $this->items = [];
+        $this->resolvedItems = [];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function count()
+    {
+        return \count($this->items);
     }
 }